Move query tests to RealmQueryTest.
